+++ src/file_analysis.h
@@ -11,6 +11,6 @@ struct file_analysis
 
 struct file_analysis *file_analysis_new();
 void file_analysis_delete(struct file_analysis *file_analysis);
-bool file_analysis_parse_line(char *line, char **file, int *char_base10, int *occurrences);
+bool file_analysis_parse_line(char *line, char **file, int *char_base10, unsigned long *occurrences);
 
 #endif
