+++ src/slicer.c
@@ -56,7 +56,7 @@ void reader_listener(void *fd_v)
             // linea completata, pronta per essere analizzata
             char *file;
             int char_int;
-            int occurrences;
+            unsigned long occurrences;
 
             if (file_analysis_parse_line(a_line, &file, &char_int, &occurrences))
             {
